fix: filter types and style[WTEL-6934](https://webitel.atlassian.net/browse/WTEL-6934)
