+++ packages/ui-datalist/src/modules/filters/classes/Filter.ts
@@ -47,30 +47,6 @@ export interface IFilter {
   set: (data: FilterData) => IFilter;
 }
 
-export interface FilterEmits {
-  'update:filter': [FilterData];
-  'delete:filter': [IFilter];
-}
-
-export interface StaticFilterEmits extends FilterEmits {
-  'add:filter': [FilterInitParams];
-}
-
-export interface DynamicFilterEmits extends FilterEmits {
-}
-
-export interface FilterProps {
-  filter: IFilter;
-  filterConfig: AnyFilterConfig;
-}
-
-export interface StaticFilterProps extends FilterProps {
-}
-
-export interface DynamicFilterProps extends FilterProps {
-  readonly?: boolean;
-}
-
 export class Filter implements IFilter {
   readonly name: FilterName;
   label: FilterLabel;
